added user sign up, connected to db
