+++ src/models/user.js
@@ -1,0 +1,32 @@
+const mongoose = require('mongoose')
+const validator = require('validator')
+
+const User = mongoose.model('User', {
+    name: {
+        type: String,
+        required: true,
+        trim: true
+    },
+    email: {
+        type: String,
+        required: true,
+        trim: true,
+        lowercase: true,
+        validate(value){
+            if (!validator.isEmail(value)) {
+                throw new Error('Email is invalid')
+            }
+        }
+    },
+    age: {
+        type: Number,
+        default: 0,
+        validate(value) {
+            if (value < 0) {
+                throw new Error('Age must be above 0')
+            }
+        }
+    }
+})
+
+module.exports = User
