+++ web-server/app.js
@@ -1,16 +1,17 @@
 const path = require('path')
 const express = require('express')
-// const basicAuth = require('basic-Auth')
 const bodyParser = require('body-parser')
+const User = require('./models/user')
+require('./db/mongoose')
 
 
 const app = express()
+const port = process.env.PORT || 3000
 app.use(bodyParser.urlencoded({extended:true}))
-
+app.use(express.json())
 
 const publicDirectoryPath = path.join(__dirname, '../')
 
-// todo - enable home.html
 
 app.use(express.static(publicDirectoryPath))
 
@@ -24,8 +25,17 @@ app.post('/home', (req, res) => {
 })
 
 
+// Creates new user - JSON POST: name, email, password
+app.post('/users', (req, res) => {
+    const user = new User(req.body)
 
+    user.save().then(() => {
+        res.send(user)
+    }).catch((e) => {
+        res.status(400).send(e) 
+    })
+})
 
-app.listen(3000, () => {
-    console.log('The server started on port 3000')
-}) 
+app.listen(port, () => {
+    console.log('Underwater-Server is up on port ' + port)
+})
